Update README.md
Built using tensorflow. To test your own images drop them into the root file of the code with the name "test" and the extension jpeg/jpg. Feel free to mess around with activation functions and optimizers. The CNN is overfitting a bit as of right now and could use more lots more data.